pyproject.toml
[project]
name = "primevul"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasets>=4.8.4",
    "duckdb>=1.5.1",
    "httpx>=0.28.1",
    "pydantic-ai-slim[anthropic]>=1.73.0",
]
